+++ src/index.js
@@ -125,6 +125,3 @@ function help() {
         )
     );
 }
-
-// TODO make all commands here
-// TODO cluster